Rename: course_syllabus and course_semester
