+++ lib/src/connector/score_connector.dart
@@ -2,7 +2,7 @@
 // @dart=2.10
 import 'package:flutter_app/debug/log/log.dart';
 import 'package:flutter_app/src/connector/ntut_connector.dart';
-import 'package:flutter_app/src/model/course/course_class_json.dart';
+import 'package:flutter_app/src/model/course/course_semester.dart';
 import 'package:flutter_app/src/model/course/course_score_json.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart';
