+++ lib/src/model/course/course_syllabus.dart
