+++ lib/src/task/course/course_semester_task.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_app/src/connector/course_connector.dart';
-import 'package:flutter_app/src/model/course/course_class_json.dart';
+import 'package:flutter_app/src/model/course/course_semester.dart';
 import 'package:flutter_app/src/r.dart';
 import 'package:get/get.dart';
 import 'package:numberpicker/numberpicker.dart';
